Adds a simple cache file containing compliant pmid/grant combo
After testing, it was found that reloading 6000+ compliant publications takes a lot of time, even if no changes are needed. There is no need to reprocess compliant publications that already have a PMCID since nothing will change for these. This change adds a simple text file containing a list of processed pmid|grantnumber combinations that do not need to be re-processed. This acts as a lookup cache during processing to avoid unneccessary re-processing.